requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
mysql-connector-python
pymysql
pydantic